+++ benchopt/benchmark.py
@@ -449,22 +449,25 @@ class Benchmark:
         if len(shell_install_scripts) > 0 or len(conda_reqs) > 0:
             check_installs += [objective]
         to_install = itertools.chain(include_datasets, include_solvers)
-        for klass in to_install:
-            reqs, scripts, hooks, missing = (
-                klass.collect(env_name=env_name, force=force, gpu=gpu)
-            )
-            # If a class is not importable but has no requirements,
-            # it might be because the requirements are specified
-            # as global ones in the Objective. Otherwise, raise a
-            # comprehensible error.
-            if missing is not None:
-                missings.append(missing)
 
-            conda_reqs += reqs
-            shell_install_scripts += scripts
-            post_install_hooks += hooks
-            if len(scripts) > 0 or len(reqs) > 0:
-                check_installs += [klass]
+        if not minimal:
+            for klass in to_install:
+                reqs, scripts, hooks, missing = (
+                    klass.collect(env_name=env_name, force=force, gpu=gpu)
+                )
+                # If a class is not importable but has no requirements,
+                # it might be because the requirements are specified
+                # as global ones in the Objective. We keep track of them
+                # to check and raise a comprehensive error after the install
+                # if it is still not importable.
+                if missing is not None:
+                    missings.append(missing)
+
+                conda_reqs += reqs
+                shell_install_scripts += scripts
+                post_install_hooks += hooks
+                if len(scripts) > 0 or len(reqs) > 0:
+                    check_installs += [klass]
         print(colorify(' done', GREEN))
 
         # Install the collected requirements